# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TalkABC 聊天交友平台后端 API 文档",
        "title": "TalkABC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/profile/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户完成所有个人资料收集后调用此接口，标记资料收集状态为已完成，允许进入应用首页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人资料"
                ],
                "summary": "完成资料收集",
                "responses": {
                    "200": {
                        "description": "标记成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "操作失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/profile/me": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户完善个人资料信息，包括昵称、性别、出生年份、身高、体重、城市、学校、职业、教育程度、星座、爱好、交友目的等\n\n**昵称安全校验规则：**\n- 长度限制：2-20个字符\n- 字符类型：仅允许中文、英文、数字、下划线、连字符\n- 敏感词过滤：禁止包含暴力、色情、政治敏感、诈骗等违规词汇\n- URL过滤：禁止包含超链接\n- HTML过滤：禁止包含HTML标签\n- XSS过滤：禁止包含XSS攻击代码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人资料"
                ],
                "summary": "完善个人信息",
                "parameters": [
                    {
                        "description": "个人信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "保存成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或昵称包含违规内容",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "保存失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/profile/preferences": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户设置理想对象的筛选条件，包括年龄范围、性别、身高范围、体重范围、教育程度、星座、爱好等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人资料"
                ],
                "summary": "设置理想对象条件",
                "parameters": [
                    {
                        "description": "理想对象条件",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "保存成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "保存失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/profile/sign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户设置个人主页的个性签名\n\n**签名安全校验规则：**\n- 长度限制：最大200字符\n- 敏感词过滤：禁止包含暴力、色情、政治敏感、诈骗等违规词汇\n- URL过滤：禁止包含 http/https/www 等超链接\n- HTML过滤：禁止包含HTML标签\n- JavaScript过滤：禁止包含脚本代码（如 \u003cscript\u003e、eval()、alert()）\n- SQL注入过滤：禁止包含SQL注入代码（如 SELECT、INSERT、OR 1=1）\n- XSS过滤：禁止包含XSS攻击代码（如 onclick=、onerror=）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人资料"
                ],
                "summary": "设置个性签名",
                "parameters": [
                    {
                        "description": "签名内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "保存成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或签名包含违规内容",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "保存失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/profile/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "检查用户是否已完成个人资料收集，用于首次登录时判断是否需要跳转到资料收集页面",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "个人资料"
                ],
                "summary": "检查资料收集状态",
                "responses": {
                    "200": {
                        "description": "data.profile_completed\": true表示已完成，false表示需要收集",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}